inventory/requirements.txt
asgiref==3.6.0
clear==2.0.0
crispy-bootstrap5==0.7
Django==4.1.6
django-crispy-forms==2.0
djongo==1.3.6
dnspython==2.3.0
mysql==0.0.3
mysql-connector-python==8.0.32
mysqlclient==2.1.1
protobuf==3.20.3
psycopg2-binary==2.9.5
pymongo==3.12.3
PyMySQL==1.0.2
pyodbc==4.0.35
python-dotenv==1.0.0
pytz==2022.7.1
sqlparse==0.2.4
tzdata==2022.7
